model tests added

+++ models/Publications_Tags.js
@@ -16,9 +16,14 @@ class Publications_Tags extends Model {
     publicationUrl /*: string */,
     tagId /*: number */
   ) /*: any */ {
+    // check publication
+    if (!publicationUrl) return new Error('no publication')
     let publicationShortId = parseurl(publicationUrl).path.substr(13)
     const publication = await Publication.byShortId(publicationShortId)
     if (!publication) return new Error('no publication')
+
+    // check tag
+    if (!tagId) return new Error('no tag')
     const tag = await Tag.byId(tagId)
     if (!tag) return new Error('no tag')
     // check if already exists
@@ -39,11 +44,17 @@ class Publications_Tags extends Model {
     publicationUrl /*: string */,
     tagId /*: string */
   ) /*: number */ {
+    // check publication
+    if (!publicationUrl) return new Error('no publication')
     let publicationShortId = parseurl(publicationUrl).path.substr(13)
     const publication = await Publication.byShortId(publicationShortId)
     if (!publication) return new Error('no publication')
+
+    // check tag
+    if (!tagId) return new Error('no tag')
     const tag = await Tag.byId(tagId)
     if (!tag) return new Error('no tag')
+
     return await Publications_Tags.query()
       .delete()
       .where({
